+++ backend/EasyMeets.Core/EasyMeets.Core.DAL/Entities/Question.cs
@@ -1,6 +1,6 @@
 namespace EasyMeets.Core.DAL.Entities;
 
-public class Question : Entity<int>
+public class Question : Entity<long>
 {
     public int AvailabilitySlotId { get; set; }
     public string Text { get; set; } = string.Empty;
